requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pdfplumber==0.9.0
pinecone
langchain==0.0.222
python-dotenv==1.0.0
langchain
langchain-community
langchain-openai
